requirements.txt
discord
requests
random
openai
python-dotenv